remote_address: step 1 proxy all hostcalls::get_property
