+++ src/lib.rs
@@ -7,6 +7,7 @@ mod glob;
 mod operation_dispatcher;
 mod policy;
 mod policy_index;
+mod property;
 mod service;
 
 #[cfg(test)]
